feat(soql): add test for common utils
